Add an option for soft errors
In case of #1222 [1], the fail fast [2] strategy adopted here appeared
to mask an underlying issue. I suspect this might also be the case for
 #1313 [3] and similar issues. Thus an option for erroring softly is
introduced. With this option, instead of throwing fast in Cypress' event
handlers, the preprocessor will emit an error message to stderr at the
end of a run.

My hope is that this will make it easier for the end-user to provide
proper reproducible examples when running into these errors.

[1] https://github.com/badeball/cypress-cucumber-preprocessor/issues/1222
[2] https://martinfowler.com/ieeeSoftware/failFast.pdf
[3] https://github.com/badeball/cypress-cucumber-preprocessor/issues/1313