Make usage formatter tests deterministic
This is done by placing strategic waits among the tests, to ensure the
desired output with respect to sorting.

The problem of deterministic sorting was something I originally thought
best could best be solved by introducing a flag in the usage formatter in
cucumber-js [1].

However, a cuople of strategic waits was much simpler than I thought it
would be and it's sufficiently deterministic. I've run the usage
formatter tests a thousand times without any flake.

This patch has the fortunate effect of removing the need for
patch-package altogether, which has been troublesome [2].

[1] ..
[2] ..